refactor: группировка пактов функционала регистрации
